refactor: Separate Job Submission into Dedicated Endpoint
- Create new `/tapis/jobs.ts` endpoint for job submission
- Remove job submission method from `/tapis/jobs/{id}.ts`
- Maintain existing API documentation and request handling
- Improve endpoint organization by separating job submission and job status retrieval
